+++ packages/ui/src/widgets/Leaderboard/Leaderboard.tsx
@@ -11,9 +11,17 @@ import {
   Content,
   Truncate,
   Alert,
+  CopyButton,
 } from '../../components';
 import { useAsyncCallback, useConfigContext, useWalletContext } from '../../hooks';
 import { TopWidget } from './TopWidget';
+
+export type TournamentImagesType = {
+  guid: string;
+  id: string;
+  mainImage: boolean;
+  path: string;
+};
 
 type TournamentType = {
   id: number;
@@ -21,6 +29,7 @@ type TournamentType = {
   subtitle: string;
   startDate: Date;
   endDate: Date;
+  images: TournamentImagesType[];
   status: 'DISABLED' | 'ENABLED';
 };
 
@@ -48,12 +57,13 @@ const LeaderboardTitle = styled(Typography)({
 });
 
 const Address = styled.div({
-  width: 105,
+  width: 'auto',
   height: 36,
   padding: '8px 16px',
   background: 'rgba(233, 204, 255, 0.1)',
   borderRadius: 4,
   fontSize: 14,
+  display: 'flex',
 });
 
 const SignUpButton = styled.button({
@@ -133,20 +143,25 @@ Be a top 3 player to win a prize"
           score={playerData?.score}
           currentScore={currentScore}
           rank={playerData?.rank}
+          tournamentImages={activeTournament?.images}
         />
       )}
       <ModalWrapper layer={`${cdnUrl}/assets/layer.svg`}>
         <RadialGradientBackGround />
         <Content>
-          {playerData && (
+          {/*TODO change address to playerData in the future*/}
+          {address && (
             <Stack direction="row" spacing="space-between">
               <LeaderboardTitle>leaderboard</LeaderboardTitle>
               <Address>
-                <Truncate text={playerData.address} maxLength={8} />
+                {/*TODO change address to playerData.address in the future*/}
+                <Truncate text={address} maxLength={8} />
+                <CopyButton />
               </Address>
             </Stack>
           )}
-          {!playerData && (
+          {/*TODO change address to playerData in the future*/}
+          {!address && (
             <Stack align="center">
               <SignUpButton onClick={handlePlayAgain}>Sign up & reveal your rank</SignUpButton>
             </Stack>